% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatRegion0.R
\name{compute_panel_regions0}
\alias{compute_panel_regions0}
\title{Compute panel regions}
\usage{
compute_panel_regions0(
  data,
  scales,
  ref_data,
  keep = NULL,
  drop = NULL,
  stamp = F
)
}
\arguments{
\item{data}{A data frame.}

\item{scales}{Scales.}

\item{ref_data}{A reference data frame.}

\item{keep}{A character vector of IDs to keep. Optional.}

\item{drop}{A character vector of IDs to drop. Optional.}

\item{stamp}{A single logical value. Optional.}
}
\value{
A data frame. If \code{stamp} is \code{FALSE}, the processed \code{ref_data} inner-joined with \code{data}.
Otherwise, the processed \code{ref_data}.
}
\description{
A short description...
}
